# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EduGest DZ API",
        "description": "API complète — plateforme SaaS de gestion scolaire pour l'Algérie.\nAuthentification via JWT Bearer (header Authorization: Bearer <token>).\nMulti-tenant : chaque requête doit inclure le header X-Tenant-ID.",
        "contact": {
            "email": "support@edugest.dz"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Serveur courant"
        },
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/api/v1/entretien/interventions": {
            "get": {
                "tags": [
                    "Entretien"
                ],
                "summary": "Liste des interventions d'entretien",
                "operationId": "80f3bc5fa071bca0fec8b6a4cd5df85f",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/TenantId"
                    },
                    {
                        "name": "statut",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "signale",
                                "en_cours",
                                "en_attente",
                                "resolu",
                                "annule"
                            ]
                        }
                    },
                    {
                        "name": "priorite",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "urgente",
                                "haute",
                                "normale",
                                "basse"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Interventions paginées",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Entretien"
                ],
                "summary": "Signaler une intervention d'entretien",
                "operationId": "6055d249cccef0dcb97f5142bfc468fe",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/TenantId"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titre",
                                    "type",
                                    "priorite"
                                ],
                                "properties": {
                                    "titre": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "panne",
                                            "degradation",
                                            "entretien_preventif",
                                            "renovation",
                                            "nettoyage",
                                            "inspection"
                                        ]
                                    },
                                    "priorite": {
                                        "type": "string",
                                        "enum": [
                                            "urgente",
                                            "haute",
                                            "normale",
                                            "basse"
                                        ]
                                    },
                                    "local_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "nullable": true
                                    },
                                    "prestataire_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "nullable": true
                                    },
                                    "cout_estime": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Intervention créée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "SuccessResponse": {
                "properties": {
                    "success": {
                        "description": "── Réponse générique ──────────────────────────────────────────────────────",
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Opération réussie"
                    },
                    "data": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
            "TenantId": {
                "name": "X-Tenant-ID",
                "in": "header",
                "description": "Identifiant unique du tenant (école/centre)",
                "required": true,
                "schema": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}